Move actions to folder
