+++ spec/quotes/actions/createQuote.test.js
@@ -1,5 +1,5 @@
 const QuotesInMemoryRepo =  require("../../../domain/quotes/infrastructure/quotesInMemoryRepo")
-const createQuote = require("../../../domain/quotes/createQuote");
+const createQuote = require("../../../domain/quotes/actions/createQuote");
 
 describe('Create quote', () => {
     it('creates a quote ', () => {
